+++ infrastructures/doctrine/Writer/ODM/MediaWriter.php
@@ -31,7 +31,7 @@ use RuntimeException;
 use Teknoo\East\Foundation\Promise\PromiseInterface;
 use Teknoo\East\Website\Doctrine\Object\Media;
 use Teknoo\East\Website\Object\MediaMetadata;
-use Teknoo\East\Website\Object\ObjectInterface;
+use Teknoo\East\Website\Contracts\ObjectInterface;
 use Teknoo\East\Website\Writer\WriterInterface;
 use Teknoo\East\Website\Writer\MediaWriter as OriginalWriter;