+++ infrastructures/symfony/Writer/UserWriter.php
@@ -28,7 +28,7 @@ namespace Teknoo\East\WebsiteBundle\Writer;
 use RuntimeException;
 use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
 use Teknoo\East\Foundation\Promise\PromiseInterface;
-use Teknoo\East\Website\Object\ObjectInterface;
+use Teknoo\East\Website\Contracts\ObjectInterface;
 use Teknoo\East\Website\Object\User as BaseUser;
 use Teknoo\East\Website\Writer\WriterInterface;
 use Teknoo\East\Website\Writer\UserWriter as UniversalWriter;