+++ src/Writer/UserWriter.php
@@ -26,7 +26,7 @@ declare(strict_types=1);
 namespace Teknoo\East\Website\Writer;
 
 use Teknoo\East\Foundation\Promise\PromiseInterface;
-use Teknoo\East\Website\Object\ObjectInterface;
+use Teknoo\East\Website\Contracts\ObjectInterface;
 use Throwable;
 
 /**